SmartTool
The SmartTool, is a length and an angle measurer application. The
program provides a movable long ruler inside the screen so that the
users can measure more length than the size of the screen. Also, when
the phone is turned on a horizontal surface, the application is able to
detect the amount of angle turned using the magnetometer of an iPhone.
So, thanks to the program your iPhone can be used both as a ruler and
as a protractor.